bot/requirements_discordpy.txt
# Alternative: Use discord.py instead of py-cord
discord.py[voice]>=2.3.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
requests>=2.31.0
